requirements.txt
gradio
google-generativeai
serpapi
python-dotenv
